Replace Flask with Quart
Replaced Flask with Quart; originally went for Sanic, but code would
have required major rewrite (and, in it's condition, I don't want to
touch it much).  While at it, added awaits in handler, and then traced
the call chain backwards, turning functions/methods to async and adding
awaits...

Originall code is a bit of spaghetti, so... hopefully I haven't
miserably failed at this.

Haven't touched tests (and not sure I want to...).